requirements.txt
alembic==1.7.7
amqp==2.6.1
aniso8601==9.0.1
attrs==21.4.0
billiard==3.6.4.0
celery==4.4.7
Cerberus==1.3.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
dataclasses-json==0.5.7
decorator==5.1.1
deepdiff==5.8.0
dnspython==2.2.1
email-validator==1.1.3
flasgger==0.9.5
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.0.3
jsonschema==4.4.0
kombu==4.6.11
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
mistune==2.0.2
mypy-extensions==0.4.3
mysqlclient==2.1.0
ordered-set==4.1.0
packaging==21.3
py==1.11.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-decouple==3.5
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
retry==0.9.2
six==1.16.0
SQLAlchemy==1.4.29
typing-inspect==0.7.1
typing_extensions==4.2.0
urllib3==1.26.9
vine==1.3.0
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.8.0

uhashring~=2.1
redis~=4.3.1
-e git+https://uE_-dVMz7ycMqRqeMcQn:uE_-dVMz7ycMqRqeMcQn@gitlab.evgcorp.net/heandend/antmedia-client.git#egg=antmedia_client
